code: Optimized the test-code structure       tested the uploading, parsing, and Q&A for short PDF and DOCX files       removed the unimplemented DOC parsing part in file_handler
