--- src/evaluation/test_questions.json
@@ -1,6 +1,0 @@
-[
-    "RISC-V架构中定义了哪几种特权级别？",
-    "mstatus寄存器中的MIE和SIE位有什么作用？",
-    "mcause寄存器如何表示异常和中断？"
-]
-
