--- src/evaluation/test_questions_short.json
@@ -1,6 +1,0 @@
-[
-    "“荡舟心许”出自谁的诗？",
-    "作者是什么时候去散步的？",
-    "作者将荷花比作什么？"
-]
-
