+++ src/server/api_server/api_client.py
@@ -187,7 +187,7 @@ def run_test():
     # asyncio.run(test_health_check())
     # asyncio.run(test_new_knowledge_base())
     asyncio.run(test_upload_files('./这是一个测试文件.txt'))
-    # asyncio.run(test_local_doc_chat())
+    asyncio.run(test_local_doc_chat())
 
 # asyncio.run(test_upload_files('./这是一个测试文件.txt'))
